UI file restructuring & bug fixing
- Split UI files for each form so that the file generated by Qt stays untouched
- Moved animation properties to the main window so that QBoxLayout suffices and inheritance is no longer required
