+++ python/tk_syntheyes/ui/main_window.py
@@ -1,198 +1,17 @@
-# -*- coding: utf-8 -*-
-
-################################################################################
-## Form generated from reading UI file 'mainrOKrvr.ui'
-##
-## Created by: Qt User Interface Compiler version 5.15.2
-##
-## WARNING! All changes made in this file will be lost when recompiling UI file!
-################################################################################
+import os
+import sys
 
 from PySide2.QtCore import *
 from PySide2.QtGui import *
 from PySide2.QtWidgets import *
 
-
-class Ui_MainWindow(object):
-    def setupUi(self, MainWindow):
-        if not MainWindow.objectName():
-            MainWindow.setObjectName(u"MainWindow")
-        MainWindow.resize(227, 417)
-        sizePolicy = QSizePolicy(QSizePolicy.Preferred, QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(MainWindow.sizePolicy().hasHeightForWidth())
-        MainWindow.setSizePolicy(sizePolicy)
-        MainWindow.setMinimumSize(QSize(0, 90))
-        MainWindow.setWindowTitle(u"SynthEyes ShotGrid")
-        MainWindow.setLocale(QLocale(QLocale.English, QLocale.Germany))
-        self.actionExit_SynthEyes = QAction(MainWindow)
-        self.actionExit_SynthEyes.setObjectName(u"actionExit_SynthEyes")
-        self.actionAlways_On_Top = QAction(MainWindow)
-        self.actionAlways_On_Top.setObjectName(u"actionAlways_On_Top")
-        self.actionAlways_On_Top.setCheckable(True)
-        self.actionAlways_On_Top.setChecked(True)
-        self.actionOpen_Console = QAction(MainWindow)
-        self.actionOpen_Console.setObjectName(u"actionOpen_Console")
-        self.centralwidget = QWidget(MainWindow)
-        self.centralwidget.setObjectName(u"centralwidget")
-        self.vertical_layout = QVBoxLayout(self.centralwidget)
-        self.vertical_layout.setSpacing(0)
-        self.vertical_layout.setObjectName(u"vertical_layout")
-        self.vertical_layout.setContentsMargins(0, 0, 0, 0)
-        self.btn_quick_select = QPushButton(self.centralwidget)
-        self.btn_quick_select.setObjectName(u"btn_quick_select")
-        sizePolicy1 = QSizePolicy(QSizePolicy.Ignored, QSizePolicy.Fixed)
-        sizePolicy1.setHorizontalStretch(0)
-        sizePolicy1.setVerticalStretch(0)
-        sizePolicy1.setHeightForWidth(self.btn_quick_select.sizePolicy().hasHeightForWidth())
-        self.btn_quick_select.setSizePolicy(sizePolicy1)
-        self.btn_quick_select.setCheckable(True)
-
-        self.vertical_layout.addWidget(self.btn_quick_select)
-
-        self.panel_split_layout = QVSplitLayout()
-        self.panel_split_layout.setSpacing(0)
-        self.panel_split_layout.setObjectName(u"panel_split_layout")
-        self.sca_quick_select = QScrollArea(self.centralwidget)
-        self.sca_quick_select.setObjectName(u"sca_quick_select")
-        sizePolicy2 = QSizePolicy(QSizePolicy.Ignored, QSizePolicy.Ignored)
-        sizePolicy2.setHorizontalStretch(0)
-        sizePolicy2.setVerticalStretch(0)
-        sizePolicy2.setHeightForWidth(self.sca_quick_select.sizePolicy().hasHeightForWidth())
-        self.sca_quick_select.setSizePolicy(sizePolicy2)
-        self.sca_quick_select.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.sca_quick_select.setWidgetResizable(True)
-        self.sca_quick_select_contents = QWidget()
-        self.sca_quick_select_contents.setObjectName(u"sca_quick_select_contents")
-        self.sca_quick_select_contents.setGeometry(QRect(0, 0, 206, 16))
-        sizePolicy.setHeightForWidth(self.sca_quick_select_contents.sizePolicy().hasHeightForWidth())
-        self.sca_quick_select_contents.setSizePolicy(sizePolicy)
-        self.verticalLayout = QVBoxLayout(self.sca_quick_select_contents)
-        self.verticalLayout.setSpacing(6)
-        self.verticalLayout.setObjectName(u"verticalLayout")
-        self.verticalLayout.setContentsMargins(4, 4, 4, 4)
-        self.verticalSpacer = QSpacerItem(0, 0, QSizePolicy.Minimum, QSizePolicy.MinimumExpanding)
-
-        self.verticalLayout.addItem(self.verticalSpacer)
-
-        self.sca_quick_select.setWidget(self.sca_quick_select_contents)
-
-        self.panel_split_layout.addWidget(self.sca_quick_select)
-
-        self.split_layout = QHSplitLayout()
-        self.split_layout.setSpacing(0)
-        self.split_layout.setObjectName(u"split_layout")
-        self.split_layout.setContentsMargins(9, 9, 9, 0)
-        self.sca_left = QScrollArea(self.centralwidget)
-        self.sca_left.setObjectName(u"sca_left")
-        sizePolicy2.setHeightForWidth(self.sca_left.sizePolicy().hasHeightForWidth())
-        self.sca_left.setSizePolicy(sizePolicy2)
-        self.sca_left.setFrameShape(QFrame.NoFrame)
-        self.sca_left.setWidgetResizable(True)
-        self.pnl_left = QWidget()
-        self.pnl_left.setObjectName(u"pnl_left")
-        self.layout_left = QVBoxLayout(self.pnl_left)
-        self.layout_left.setSpacing(0)
-        self.layout_left.setObjectName(u"layout_left")
-        self.layout_left.setContentsMargins(0, 0, 0, 0)
-        self.sca_left.setWidget(self.pnl_left)
-
-        self.split_layout.addWidget(self.sca_left)
-
-        self.sca_right = QScrollArea(self.centralwidget)
-        self.sca_right.setObjectName(u"sca_right")
-        sizePolicy2.setHeightForWidth(self.sca_right.sizePolicy().hasHeightForWidth())
-        self.sca_right.setSizePolicy(sizePolicy2)
-        self.sca_right.setFrameShape(QFrame.NoFrame)
-        self.sca_right.setWidgetResizable(True)
-        self.pnl_right = QWidget()
-        self.pnl_right.setObjectName(u"pnl_right")
-        self.layout_right = QVBoxLayout(self.pnl_right)
-        self.layout_right.setSpacing(0)
-        self.layout_right.setObjectName(u"layout_right")
-        self.layout_right.setContentsMargins(0, 0, 0, 0)
-        self.sca_right.setWidget(self.pnl_right)
-
-        self.split_layout.addWidget(self.sca_right)
-
-        self.split_layout.setStretch(0, 1000)
-
-        self.panel_split_layout.addLayout(self.split_layout)
-
-        self.panel_split_layout.setStretch(1, 1000)
-
-        self.vertical_layout.addLayout(self.panel_split_layout)
-
-        MainWindow.setCentralWidget(self.centralwidget)
-        self.menubar = QMenuBar(MainWindow)
-        self.menubar.setObjectName(u"menubar")
-        self.menubar.setGeometry(QRect(0, 0, 227, 21))
-        self.menuFile = QMenu(self.menubar)
-        self.menuFile.setObjectName(u"menuFile")
-        MainWindow.setMenuBar(self.menubar)
-        self.statusBar = QStatusBar(MainWindow)
-        self.statusBar.setObjectName(u"statusBar")
-        self.statusBar.setSizeGripEnabled(False)
-        MainWindow.setStatusBar(self.statusBar)
-
-        self.menubar.addAction(self.menuFile.menuAction())
-        self.menuFile.addAction(self.actionOpen_Console)
-        self.menuFile.addAction(self.actionExit_SynthEyes)
-
-        self.retranslateUi(MainWindow)
-
-        QMetaObject.connectSlotsByName(MainWindow)
-    # setupUi
-
-    def retranslateUi(self, MainWindow):
-        self.actionExit_SynthEyes.setText(QCoreApplication.translate("MainWindow", u"Exit SynthEyes", None))
-        self.actionAlways_On_Top.setText(QCoreApplication.translate("MainWindow", u"Always On Top", None))
-        self.actionOpen_Console.setText(QCoreApplication.translate("MainWindow", u"Open Console", None))
-        self.btn_quick_select.setText("")
-        self.menuFile.setTitle(QCoreApplication.translate("MainWindow", u"File", None))
-        pass
-    # retranslateUi
-
-################################################################################
-
-class SplitProperty(object):
-    @property
-    def stretch_steps(self):
-        return self.stretch(0) + self.stretch(1)
-
-    def get_split_factor(self):
-        stretch_steps = self.stretch_steps
-        if stretch_steps:
-            return self.stretch(0) / stretch_steps
-        return 0.5
-
-    def set_split_factor(self, factor: float):
-        """Set the stretch factors to resemble the given factor in range [0,1]"""
-        stretch_steps = self.stretch_steps
-        self.setStretch(0, round(factor * stretch_steps))
-        self.setStretch(1, round((1 - factor) * stretch_steps))
-    
-    split_factor = Property(float, get_split_factor, set_split_factor)
-
-class QHSplitLayout(QHBoxLayout, SplitProperty):
-    def __init__(self, parent=None):
-        super(QHSplitLayout, self).__init__(parent)    
-
-class QVSplitLayout(QVBoxLayout, SplitProperty):
-    def __init__(self, parent=None):
-        super(QVSplitLayout, self).__init__(parent)
-
-################################################################################
-import os
-import sys
-
 if __name__ == "__main__":
     sys.path.append(os.path.join(os.path.dirname(__file__), "..", "..", ".."))
 
-sys.path.append(os.path.join(os.path.dirname(__file__)))
-from app_command import AppCommand
-from base_panel import BasePanel, SynthEyesPanel
+sys.path.append(os.path.dirname(__file__))
+from ui_main_window import Ui_MainWindow
+from base_panel import BasePanel
+from tk_syntheyes.app_command import AppCommand
 from engine import SynthEyesEngine
 
 
@@ -242,18 +61,53 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         # Disable standby panels just in case
         self.pnl_right.setEnabled = False
 
-        # Setup animations
-        self._anim_panel_transition = QPropertyAnimation(self.split_layout, b"split_factor", self)
+        ### Setup animations ###
+        self._anim_panel_transition = QPropertyAnimation(self, b"panel_split_factor", self)
         self._anim_panel_transition.finished.connect(self._switch_panel_finished)
-        self._anim_panel_quick_select_transition = QPropertyAnimation(self.panel_split_layout, b"split_factor", self)
+
+        self._anim_panel_quick_select_transition = QPropertyAnimation(self, b"main_split_factor", self)
         self._anim_panel_quick_select_transition.setEasingCurve(QEasingCurve.OutExpo)
         self._anim_panel_quick_select_transition.finished.connect(self._switch_quick_select_finished)
+        ########################
 
         # Setup menu actions
         self.actionExit_SynthEyes.triggered.connect(self.exit)
         self.actionOpen_Console.triggered.connect(self.open_logging_console)
 
     
+    def get_layout_stretch_steps(self, layout: QBoxLayout):
+        """Return the total amount of stretch of the given layout."""
+        return layout.stretch(0) + layout.stretch(1)
+
+    def get_split_factor(self, layout: QBoxLayout):
+        """Get the stretch ratio of the layout stretch values as a factor in range [0,1]"""
+        stretch_steps = self.get_layout_stretch_steps(layout)
+        if stretch_steps:
+            return layout.stretch(0) / stretch_steps
+        return 0.5
+
+    def set_split_factor(self, layout: QBoxLayout, factor: float):
+        """Set the stretch values of the layout to resemble the given factor in range [0,1]"""
+        stretch_steps = self.get_layout_stretch_steps(layout)
+        layout.setStretch(0, round(factor * stretch_steps))
+        layout.setStretch(1, round((1 - factor) * stretch_steps))
+    
+    def set_main_panel_split_factor(self, factor: float):
+        self.set_split_factor(self.main_split_layout, factor)
+
+    def get_main_panel_split_factor(self):
+        return self.get_split_factor(self.main_split_layout)
+    
+    def set_split_panel_split_factor(self, factor: float):
+        self.set_split_factor(self.panel_split_layout, factor)
+
+    def get_split_panel_split_factor(self):
+        return self.get_split_factor(self.panel_split_layout)
+
+    main_split_factor = Property(float, get_main_panel_split_factor, set_main_panel_split_factor)
+    panel_split_factor = Property(float, get_split_panel_split_factor, set_split_panel_split_factor)
+
+
     def _init_panel(self, panel_type: type, name: str, parent_panel: QWidget, visible=False, enabled=False, add_to_quick_select=True):
         if parent_panel:
             if not hasattr(parent_panel, "sub_panels"):
@@ -278,7 +132,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
         if parent_panel:
             parent_panel.sub_panels[name] = panel
-            if panel_type is SynthEyesPanel or issubclass(panel_type, SynthEyesPanel):
+            if panel_type is BasePanel or issubclass(panel_type, BasePanel):
                 self._link_panel(panel.make_sub_panel(parent_panel), parent_panel)
 
         return panel
@@ -409,11 +263,11 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
         self._active_panel_depth += 1 if left_to_right else -1
         if self._left_panel_active and not left_to_right:
-            self.pnl_right.layout().addWidget(self.pnl_left.takeAt(0).widget())
-            self.split_layout.split_factor = 0
+            self.pnl_right.layout().addWidget(self.pnl_left.layout().takeAt(0).widget())
+            self.set_split_panel_split_factor(0)
         elif not self._left_panel_active and left_to_right:
-            self.pnl_left.layout().addWidget(self.pnl_right.takeAt(0).widget())
-            self.split_layout.split_factor = 1
+            self.pnl_left.layout().addWidget(self.pnl_right.layout().takeAt(0).widget())
+            self.set_split_panel_split_factor(1)
 
         # Switch panels if direction is reversed
         if left_to_right:
@@ -428,8 +282,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         target_panel.setEnabled(True)
         panel_to.layout().addWidget(target_panel)
         
-        split: QHSplitLayout = self.split_layout
-        split_factor = split.split_factor
+        split_factor = self.get_split_panel_split_factor()
         duration = 200 * (1 - split_factor if split_factor < end_value else split_factor)
 
         self._left_panel_active = not left_to_right
@@ -467,7 +320,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         quick_select = self.btn_quick_select.isChecked()
 
         # If panel is not fully opened/closed shrink duration respectively
-        split_factor = self.panel_split_layout.split_factor
+        split_factor = self.get_main_panel_split_factor()
         duration = 300 * (1 - split_factor if split_factor < quick_select else split_factor)
 
         self._anim_panel_quick_select_transition.setEndValue(quick_select)
